fixing cli; cli util pckg ; tweak names and messages on server;

+++ server/dto/gameResultDto.go
@@ -9,7 +9,7 @@ type GameResultDto struct {
 	Id               string              `json:"id"`
 	Username         string              `json:"username"`
 	QuestionsResults []QuestionResultDto `json:"questionsResults"`
-	ScorePercentage  int                 `json:"scorePercentage"`
+	ScorePercentage  float64             `json:"scorePercentage"`
 	CreatedDate      time.Time           `json:"createdDate"`
 	CompletedDate    time.Time           `json:"completedDate"`
 	PercentileScore  int                 `json:"percentileScore"`
